+++ R/get_prism_monthlys.R
@@ -30,16 +30,6 @@
 #'   keepZip = TRUE
 #' )
 #' 
-#' # Get pre-1981 data (resolution applies to both pre and post-1981 data)
-#' get_prism_monthlys(
-#'   type = "tmax",
-#'   years = 1975,
-#'   mon = 7,
-#'   resolution = "4km",
-#'   keep_pre81_months = FALSE,
-#'   keepZip = FALSE
-#' )
-#' 
 #' # will fail - invalid resolution:
 #' get_prism_monthlys(
 #'   type = "ppt",
@@ -54,7 +44,7 @@
 #' 
 #' @export
 get_prism_monthlys <- function(type, years, mon = 1:12, keepZip = TRUE,
-                               keep_pre81_months = TRUE, service = NULL, 
+                               keep_pre81_months = NULL, service = NULL, 
                                resolution = "4km")
 {
   ### parameter and error handling
@@ -87,87 +77,40 @@ get_prism_monthlys <- function(type, years, mon = 1:12, keepZip = TRUE,
     stop("'resolution' must be '4km' or '800m'. See ?get_prism_monthlys for details.")
   }
   
-  pre_1981 <- years[years<1981]
-  post_1981 <- years[years>=1981]
-  uris_pre81 <- vector()
-  uris_post81 <- vector()
+  if (!is.null(keep_pre81_months)) {
+    warning('`keep_pre81_months` is deprecated and no longer has any effect. It will be removed in a future release.')
+  }
+  
+  uris <- vector()
   
   if (is.null(service)) {
 	  service <- "http://services.nacse.org/prism/data/public/4km"
   }
-  
-  if (length(pre_1981)) {
-    # uris_pre81 <- gen_prism_url(pre_1981, type, service)
-    uris_pre81 <- gen_prism_url(pre_1981, type, resolution)
-  }
 
-  if (length(post_1981)) {  
-    uri_dates_post81 <- apply(
-      expand.grid(post_1981,mon_to_string(mon)), 
-      1, 
+  if (length(years)) {
+    uri_dates <- apply(
+      expand.grid(years, mon_to_string(mon)),
+      1,
       function(x) {paste(x[1], x[2], sep="")}
     )
-    
+
     # uris_post81 <- gen_prism_url(uri_dates_post81, type, service)
-    uris_post81 <- gen_prism_url(uri_dates_post81, type, resolution)
+    uris <- gen_prism_url(uri_dates, type, resolution)
   }
     
   download_pb <- txtProgressBar(
     min = 0,
-    max = length(uris_post81) + length(uris_pre81),
+    max = length(uris),
     style = 3
   )
- 
-  counter <- 0
 
-  ### Handle post 1981 data
-  if(length(uris_post81) > 0){    
-      for(i in seq_along(uris_post81)){
-        prism_webservice(uris_post81[i],keepZip)
+  ### Handle all data
+  if(length(uris) > 0){    
+      for(i in seq_along(uris)){
+        prism_webservice(uris[i],keepZip)
         setTxtProgressBar(download_pb, i)
     }
   }
-    
-  counter <- length(uris_post81)+1
-
-  ### Handle pre 1981 files
-  if (length(uris_pre81) > 0) {
-  
-    pre_files <- c()
-    for (j in seq_along(uris_pre81)) {
-      tmp <- prism_webservice(
-        uris_pre81[j], 
-        keepZip, 
-        returnName = TRUE,
-        pre81_months = mon
-      )
-      if (!is.null(tmp)) {
-        pre_files <- c(pre_files, tmp)
-      }
-      setTxtProgressBar(download_pb, counter) 
-      counter <- counter + 1
-    }
-    
-    # Process pre 1981 files (unzip and keep monthly data)
-    if (length(pre_files) > 0) {
-      pre_files <- unlist(strsplit(pre_files,"\\."))
-      pre_files <- pre_files[seq(1, length(pre_files), by = 2)]
-      
-      for (k in seq_along(pre_files)) {
-        yr <- regmatches(pre_files[k], regexpr('[0-9]{4}', pre_files[k]))
-        
-        if (keep_pre81_months) {
-          monstr <- c(mon_to_string(1:12), "")
-        } else {
-          monstr <- mon_to_string(mon)
-        }
-        
-        to_split <- stringr::str_replace(pre_files[k], "_all", monstr)
-        
-        process_zip(pre_files[k], to_split)
-      }
-    }
-  }
-
+ 
   close(download_pb)
 }